+++ src/pages/about.mdx
@@ -1,7 +1,7 @@
 import { SeoHeaders } from "components/ui/SeoHeaders"
-import { Layout } from "components/Layout/Layout"
-import { Wrapper } from "components/Wrapper/Wrapper"
-import { Wysiwyg } from "components/Wysiwyg/Wysiwyg"
+import { Layout } from "components/ui/Layout/Layout"
+import { Wrapper } from "components/ui/Wrapper/Wrapper"
+import { Wysiwyg } from "components/ui/Wysiwyg/Wysiwyg"
 
 # About me
