+++ src/pages/index.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import { InferGetStaticPropsType, GetStaticProps } from "next"
 import { SeoHeaders } from "components/ui/SeoHeaders"
-import { Layout } from "components/Layout/Layout"
-import { Wrapper } from "components/Wrapper/Wrapper"
+import { Layout } from "components/ui/Layout/Layout"
+import { Wrapper } from "components/ui/Wrapper/Wrapper"
 import { ArticleCard } from "components/model/post/ArticleCard/ArticleCard"
 import { ArticleCardWrapper } from "components/model/post/ArticleCardWrapper/ArticleCardWrapper"
 import { PostData } from "../components/interface/post-data"
