+++ src/pages/posts/[slug].tsx
@@ -2,9 +2,9 @@ import { GetStaticPaths, GetStaticProps, InferGetStaticPropsType } from "next"
 import markdown from "markdown-it"
 import { getHighlighter } from "shiki"
 import { isSupportedLanguage } from "utils/isSupportedLanguage"
-import { Layout } from "components/Layout/Layout"
-import { Wrapper } from "components/Wrapper/Wrapper"
-import { Wysiwyg } from "components/Wysiwyg/Wysiwyg"
+import { Layout } from "components/ui/Layout/Layout"
+import { Wrapper } from "components/ui/Wrapper/Wrapper"
+import { Wysiwyg } from "components/ui/Wysiwyg/Wysiwyg"
 import { BlogPostRepository } from "infra/blog-post-repository"
 import { PostData } from "components/interface/post-data"
 import { getSlug } from "utils/getSlug"
